Renew the scheduler interface
I'm going to implement a new module called "sched". It needs access to
the global task lock, and the head of the linked list could be useful.
If some other things are needed, the structure can be extended.

The new module will override the simple default scheduler by updating
the yield function pointer. This design should make the real scheduler
loosely coupled with the task switching mechanism.