Simplify task_yield function
The if statement in the task_yield function did not look good. Now the
yield function pointer contains a valid value before loading the IDT.
The Clang compiler produced following assembly code (32- and 64-bit):

_task_yield:
        mov eax, dword [_kernel]
        jmp dword [eax+0x124]

task_yield:
        mov rax, qword [kernel]
        jmp qword [rax+0x198]

This optimization is called "tail-call elimination", according to the
Wikipedia. It is theoretically possible that a compiler could generate
instructions that access the data structure member in a non-atomic way
and this detail has bothered me.