Improve init messages
If there is no supported video framebuffer, the boot loader prints a
message using BIOS or UEFI services. The operating system will be in
a "headless computer" mode. I have to make sure that this mode will
be tested more often. When selecting the video mode, the special key
sequence ALT + F1 will switch to "text mode", which can be used for
testing this "without-a-monitor" mode.