Fix mm_bitmap allocation
When testing the system, I found out that something is wrong somewhere
and page faults can happen when reading or writing to allocated pages.

The bitmap variables could typically be something like this:

        mm_bitmap:      FFFFFF81 00024000  <- BUG
        mm_bitmap_size: 00000000 00024000

A very obvious and serious bug above. The correct values would be:

        mm_bitmap:      FFFFFF81 00000000  <- CORRECT
        mm_bitmap_size: 00000000 00024000

I even checked these when I wrote the code, but I totally overlooked
the obvious inconsistency. Luckily, it was found early enough!