Introduce fb_enter and fb_leave functions
There were no problems because the console.c code was the only user of
the standard framebuffer (the buffer with a defined pixel format). Now
some other code could also write to the same buffer.

        fb_enter();
        do_some_drawing();
        fb_leave();

The last function, fb_leave, will render the content from the standard
buffer to the physical framebuffer.