Change global flag policy
The standard framebuffer (the buffer with a defined pixel format) uses
the D bits and copies "modified" memory pages to the real physical
framebuffer. However, on multiprocessor systems it is important that
loading the CR3 register invalidates the Translation-Lookaside Buffer
for these special pages. One task at a time is modifying the buffer,
and monitoring the D bits, but task switching can transfer control to
a different CPU. Whenever that happens, it always involves a CR3 load.
This is somewhat complex, but everything should be correct now.

As another change, I postponed setting the pg_ready variable so that
application processors would not use the paging structures until the
timer interrupt handler (I/O APIC) has been patched.