Renew gcc.sh
The cross compiler (./external/bin/x86_64-w64-mingw32-gcc) is the
primary compiler. Many host systems provide "gcc-mingw-w64" packages
so it is not strictly necessary to build it from source code.

I am not a bash programming expert, but it is worth mentioning that
"set -e" is very relevant. If the compiler fails, the return value
will terminate the Makefile process.