Use bash consistently
All other *.sh scripts are called in the same way. The file extension
is a little bit misleading, but it is like a common convention.

These scripts could be shell-independent, but I am almost sure that
building the external tools would be difficult if "bash" were not
available. The "set -e" command is important in my scripts and I have
not checked are there other implementations that support it.