Use latest versions
If building the tools from source code, use the following versions:

    NASM - The Netwide Assembler - Version 2.15.05

    GNU Binutils - Version 2.37

    GCC (The GNU Compiler Collection) - Version 11.2.0

It is not necessary to build anything if the host operating system
already has the necessary tools. Modern Linux distributions could have
packages "binutils-mingw-w64, gcc-mingw-w64, nasm" available in their
native package manager. However, the versions could be much older but
the Dancy source code does not rely on newest compiler or assembler
features.

Actual compiler or assembler bugs are a different matter. But plain C
code and simple assembly (as far as translation is concerned) should
work with older or newer versions, and I think it is very unlikely
that my code would trigger bugs that were missed when testing stable
releases.