Modify compiler options
This is not my favorite commit. I don't know exactly what are the most
optimal options so that "clang" and "gcc" both support them.

Some of the options may be redundant, but I tried to be as specific as
possible. At least some major projects seem to use similar options,
although not identical, and I think compiler developers most probably
consider the "what is enabled by default" policy carefully.