Add event_wait_array function
It is useful to be able to wait for many events. The main idea is from
the "EFI_BOOT_SERVICES.WaitForEvent()" service. However, my interface
is different and does not need any special "timer events" if callers
want to wait for a specified time.

If "uint16_t milliseconds" is zero, then this function does not sleep.
The other special value, 0xFFFF, means that the function simply waits
until at least one event is signaled.

Any non-negative return value is an array index for the first signaled
event. The return value -1 means that the timeout has expired. Other
negative values are unspecified errors.