Develop keyboard.c
I wanted to use "scan code set 2" and disable the translation in the
PS/2 controller. However, it seems that these devices are emulated
more often than not, even though I disabled "USB Legacy Support". It
is most probably safest to use the "default settings", so I decided
to continue using the controller translation, i.e. "scan code set 1".

This file contains functions that are not thread-safe, and it has been
designed so. The "ps2_task" and its initialization are the only ones
that are allowed call these functions.