Add vfs.at
This is a very essential module. The virtual file system is said to be
an abstract layer on top of a more concrete file system, and it allows
applications to access different file systems in a uniform way. This
kind of description is from Wikipedia:

        https://en.wikipedia.org/wiki/Virtual_file_system

My viewpoint is a little bit different. There will be file systems at
some point, but initially I'm more interested in "special files", e.g.
"/dev/zero", "/dev/null", and "/dev/ttyS*" devices. The idea of having
a uniform interface to do "open, close, read, write, ..." operations
is part of the Unix-like world. I'm trying to learn the most practical
and useful things from that world.

Going too much in that direction may be detrimental to innovation.