Update scripts
The source code hierarchy is a little bit better. Now at the end of
the year I have started planning a new directory for user space. The
name should begin with the letter "a" because I would like it to be
the first directory:

        a...?
        boot
        common
        include
        kernel
        scripts
        share
        tools

That new directory could be like a "sysroot" for user space and it can
become large. After this, I think there should be no more directories
here in the root directory of the source code repository.