Introduce Arctic Dancy
All the user space code will be placed into this directory. The name
"etymology" comes from the "Arctic Circle". The memory layout of the
user space tasks is basically splitted into two parts. The low part,
so called kernel space, is relatively small and is at the bottom. It
is shared between all tasks. The high part is called user space, but
on 64-bit CPUs there is another kernel space area at the top.

Anyway, the user space is above the low kernel, and this limit is the
"Arctic Circle" in this background story. The user space has usually
memory mappings quite sparsely, so this "north side" should somehow
describe this aspect too when comparing it to the globe.