Develop vfs.c
I don't have good rationale for all design decisions, and this should
be considered as a scratch pad. However, I'm sure that it will evolve
to a good system, but occasionally I might get lost on the wrong track
and the solutions have to be rethought later.

Everything is totally unstable in this module and its interfaces are
still under construction.