Discard symbolic links
This is a difficult decision, but I will try to get along without any
symbolic links. I have been studying this subject for some time and I
did not even know how "ambiguous" the handling of working directories
and paths with "symbolic links to directories" may end up being.

If I change the current working directory:

        cd /a/b/c

Then, if I do something like this:

        cd ../c
        ls ../c

The only reasonable current working directory should be "/a/b/c" and
the file list should be from that same directory. I am totally wrong:
with symbolic links this is ambiguous. To make this whole thing even
more confusing, some shells try to hide the "problem" and behave in a
different (more logical) way when it comes to the "cd" command.