Add unlink functions
Basically "unlinking" is like "removing" a file in my current design.
I don't know if there will ever be multiple hard links pointing to the
same node. Time will tell, and I want to keep that option available.

Anyway, "unlinking" is a very familiar concept in the Unix-world and
the term should not be misleading at the kernel level interface.