Introduce epoch.c
Finally I have this code implemented. The epoch here means the number
of seconds that have elapsed since Thursday 1 January 1970 00:00:00.
The value type in my implementation is "unsigned long long", and its
range is [0, 253402300799].

The maximum value is December 31 9999 23:59:59 because the next value
would have a 5-digit year. However, all values above the maximum will
use the "saturation arithmetic" and behave in a defined way.