Use vfs_stat structure
The resolution of all the timestamps in the FAT filesystem is not very
high. In the previous commit message, I questioned the rationality of
the "last access time". What if that had not existed and longer names
for files were possible and/or the resolution of "last modidified" was
higher. Assuming that the directory record would still be 32 bytes.

One other big problem with the FAT filesystem is the file size limit.
The volume size limit is not a problem because file allocation tables
would not scale well with extremely large volumes. Of course, it goes
without saying that its performance with large files would not be very
good either, but for archiving and when considering the typical file
sizes nowadays, the file size limit is the most relevant problem.