Develop renaming
It is not an error to do vfs_rename("existing_file", "existing_file").
Now there is a new error, DE_RENAME, and it means that renaming can't
be done if "old_name" and "new_name" use different mount points.

Basically this means that callers must take the DE_RENAME error into
account anyway because of the mount point restriction, which will be
a desired feature. The FAT driver slightly "misuses" this error code
and avoids implementing the real file/directory moving logic.