Check path components
My existing FAT driver creates the parent directories automatically.
However, this is not very compatible with other operating systems so
I decided to "disable" it.

Writing a brand new file system driver is going to be simpler because
the VFS layer will guide the implementation. This fat_io.c file is a
bit messy and inelegant, but I still try to focus on correctness.