Improve vfs_mount function
This will solve the "readdir" problem that I mentioned in the previous
commit message. In short, the target (e.g. file or dir) must exist and
have the same node type.

The same mount cannot be used many times and the "owner components" of
the path cannot be used either.