Add epoch.c
This is basically the interface for using the real-time clock. It is
important to understand that my kernel itself does not need the "real"
date and time. The only reliable time notion is the kernel timer when
it comes to the events and the like.

The callers must take into account that return values may be less than
what have been previously read.