Implement floppy.c
Now I have a floppy disk driver, but unfortunately I did not have a
chance to test it on real hardware yet. It should have enough features
for now, e.g. "1440 KB" and "720 KB" 3.5-inch formats are detected in
an automatic way. Please note that having absolute guarantees is very
difficult, and other OSs have the same issues when it comes to this.

I will do a lot of tests on real hardware later. I really want to make
this stable and reliable.