Simplify fat_io_add function
There are two important reasons for this modification. First, it makes
code simpler, and there is only a single "fat_create" function call in
this compilation unit.

Second, the code path for handling the "media changed" event is always
used at least once. This reason is more important than the first one.
In practice, that kind of code will be tested more thoroughly.