Develop fat_io.c
The instance member must be read after calling the enter_fat function.
When creating the first node of the file system, it is not actually a
good idea to call the enter_fat function, so that code was removed.

Now it is possible to create the first node without the media and rely
on the "media changed" error code, which should be handled anyway.