Add fdn.c
The device nodes for floppy disks will just call functions from the
floppy.c compilation unit. However, caching will be a very important
feature in the future. The memory manager should have some kind of
way to call the fdn_free_cache function if the system is running out
of memory, but I have not implemented that functionality yet.