Mount floppy disks
My idea is to try to make this as simple as possible for the users. It
is not necessary to manually "mount" floppies, but the "/mnt/floppy_?"
directories can be directly used. If there is a file opened, and there
is a "media changed" event, then that specific node would still exist
until it is released but all actions would result in a "media changed"
error code.

The root directory of the floppy is special and its node can exist all
the time, whether or not there is media. The "readdir" can be used to
get started using a floppy disk, and no I/O happens before that. Also,
the file can be opened directly if the name is known.