Use uint32_t for readdir
The "offset" name was a bit misleading, because the value is an index.
Having a 64-bit unsigned integer was an exaggeration. If 32 bits were
not enough as an index in a single directory, then it would make sense
to re-evaluate other types too.

This uint32_t is better balanced with uint64_t when we are talking
about "maximum number of files" versus "file sizes".