Change readdir strategy
The first byte of the directory entry is special. 0xE5 means that the
entry is free. 0x00 means that the entry is free and all other entries
in the same cluster are also free. 0x05 is special in some character
sets, but I don't have any support for other than ASCII characters in
the FAT file system.

I have been worrying about the 0x00 case. Good drivers make sure that
all the first bytes of the directory entries are literally 0x00 in the
rest of the cluster (which should be the last one in the chain). There
should be no issues if readdir passed the first entry that has 0x00.

However, if I force the driver to enumerate the entries in a logical
order, then there is no access to the "possibly uninitialized" space.
Theoretical concerns are not valid enough by themselves, so figured
out that maybe it would be a good idea anyway to force this logical
reading order. For example, it should never make much sense to check
the fifth directory entry without first checking the fourth entry. It
is allowed to read the previous entries again, however.

The directory entries in the FAT file system are not in any particular
order so they should definitely be checked sequentially in all common
use cases I can imagine.