Optimize db_read function
Assume that no other files can be found in the same database if there
are enough "padding bytes" returned by the bitarray_fetch function.
With the current "database groups", this optimization is relevant and
could avoid "hundreds of thousands" bitarray_fetch calls.

On modern hardware, it is questionable whether it actually matters or
not, but having such a mindless loop is not a good coding convention.