Develop ide_ctrl.c
There are a lot of details that need to be taken into account. I have
not tested this extensively, but this code should have enough features
so that I'm able to start the testing process. Most probably this code
will be used a lot, and making it stable and reliable is important.