Implement coff.c
The coff_load_executable function is very essential. It is a bug if
this function causes a page-fault exception. The executable file can
be erronenous and still be handled in a controlled way. If it is good
enough for the loader, it is valid for the kernel as well.

When jumping to the start function, in user space, its behavior can
be totally undefined but that should be safe.