Implement n_append function
If using this function, all writes to the file will be forced to the
current end-of-file. Now it is possible to support the O_APPEND flag.
There are some devices, e.g. /dev/null, that do not have an offset so
I added a default function, which simply calls the n_write function
with an (uint64_t)-1 offset.