Add dancy_syscall_spawn function
Mapping the file descriptors is possible if using fd_count and fd_map
arguments. The return value is pid_t of the new process if there are
no errors. In that normal case, the caller knows that the executable
image has been accepted and there are enough resources to start it.