Introduce file_getdents function
This function builds a buffer that can be accessed like standard argv
or envp arguments. There is an array of pointers to strings, and the
last one is a NULL pointer. Everything is in the same buffer.

To make this interface more future-proof, there is a flags argument.
It must be zero at the moment. In the future, the structure could be
more complicated if other directory entry information is needed.