Develop file_getdents function
The count argument is for input and output. It could be useful for the
caller to know how many directory entries were read. The other change
is much more subtle. If the "end of directory" has been reached once,
now it is guaranteed that subsequent calls have the same condition.