Implement directory functions
For most directories, the readdir function calls the getdents system
function twice. First the entries are read into the buffer, and then
the second call returns an "end of directory" condition. If there are
too many entries in a directory, then the getdents system function is
called as many times as needed.