Update db.c
This is the most interesting case. The -Wanalyzer-malloc-leak feature
of GCC complained about this, and the warning is somewhat valid.

If the malloc function returned NULL, it would not make much sense to
free it, although free(NULL) is definitely valid.

If it returned a valid pointer, which cannot be dereferenced, the it
would make some sense to free it. It depends on the implementation.