Develop mouse.c
The "/dev/dancy-mouse" device is created if other initializations work
as intended. If there are no open file descriptors for the device, the
handler function will automatically keep the history clean.

When someone opens the device, all data read from it are recent. When
it comes to mouse events, I think this approach is very reasonable.