Develop ioctl.c
The assembly function takes care of this conversion. Pointers in user
space are definitely not valid if their integer values are above the
LLONG_MAX limit. Previous code set them to NULL (zero on this system)
but now the new assembly function returns LLONG_MAX.

This detail might be very important because NULL is a special case.