Improve pg_check_user function
The stack area in user space is dynamically allocated when it is read
or written. When doing this check itself, it is interpreted as reading
and then real physical pages are allocated when needed.

If the pg_check_user function does not return an error, it means that
the whole area requested for checking is pointing to real memory.