Fix remainder bugs
A very annoying mistake in thinking has happened in the past. If there
are ms64 values above what uint32_t variables can hold, this assertion
is totally _wrong_:

        extern uint64_t ms64;
        uint32_t ms32 = (uint32_t)ms64;

        assert((ms64 % 1000) == (ms32 % 1000));  /* BUG! */

The reason for having these "tricks" is the missing "___udivmoddi4"
function. The new solution which uses the volatile pointer is correct
but not very elegant.