Implement reboot.c
Today this operating system got a very nice feature. If users run the
'reboot' or 'poweroff' programs, the computer behaves in the expected
way. ACPICA can be built using the following 'make' command:

    make ACPICA_O32_AT=./o32/acpica.at ACPICA_O64_AT=./o64/acpica.at

The computer will turn off for real when having ACPICA. The system is
very usable even without it, and I would like to keep it that way.