Develop fat_io.c
Case sensitivity is a big challenge in this file system. However, now
it has been clearly defined: filenames cannot contain non-printable
ASCII characters or uppercase letters. There are more restrictions in
the actual driver (including reserved names).

The driver supports 8.3 filenames only. The names are uppercase on the
disk, but are shown lowercase for the operating system.