Add init/video/_driver.c
These new functions do nothing at the moment. However, it is possible
to implement "native video drivers" that can set up proper video modes
for the operating system. The boot firmware does not always support
the native resolution of the display.

If this works well, drivers can be placed in the init/video directory.