Develop ps2/8042.c
This "IA-PC Boot Architecture Flags" has been an issue when trying to
detect whether it is safe to initialize the "8042 Controller" or not.

I made two laptops usable after removing this check and the embedded
keyboards and mice work now. The "USB Legacy Support" features have
been disabled already, but maybe there is some other emulation in the
background.