Develop xhci.c
It is not totally clear to me whether memset and memcpy functions can
have a NULL pointer if the size parameter is zero. This is mostly
theoretical because my assembly functions behave in a defined way. It
may still be relevant if compilers do optimizations or use inlined
standard functions.