Develop USB
Now the HID driver configures the endpoints. There is no need for the
report descriptor if using the boot interface subclass, but I read it
anyway. At least my current real USB keyboard works well and is very
fast and responsive! I can _use_ the OS on this desktop computer.