Develop reboot.c
This is a major improvement! All the tasks are forced to use a single
processor when the system is going down.

After a short while, application processors will end up finding some
tasks to run next but none of the existing tasks are allowed. It is
not possible to create new tasks either.

At that point, it is much safer to run procedures (base/runlevel.c)
that are needed for actual rebooting or powering off.