Finalize reboot.c
I started to make changes to the reboot procedures yesterday, and it
turned out to be more complicated than what I thought. As can be seen,
I have made many commits recently, but this should be the final one in
this change series.

Each application processor is forced to run a halt task. After that,
the actual procedures for rebooting or powering off are called.